adds the ability to parse a pdf through the Pdf::Parser.parse method
